src/backend/python/requirements.txt
openai
ollama
requests
beautifulsoup4
python-dotenv
ipython
rich